# LUCENE-10627: Using ByteBuffersDataInput reduce memory copy on compressing data