--- a/lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/compressing/CompressionMode.java
+++ b/lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/compressing/CompressionMode.java
@@ -14,14 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.codecs.compressing;
+package org.apache.lucene.backward_codecs.compressing;
 
 import java.io.IOException;
 import java.util.zip.DataFormatException;
 import java.util.zip.Deflater;
 import java.util.zip.Inflater;
 import org.apache.lucene.index.CorruptIndexException;
-import org.apache.lucene.store.ByteBuffersDataInput;
 import org.apache.lucene.store.DataInput;
 import org.apache.lucene.store.DataOutput;
 import org.apache.lucene.util.ArrayUtil;
@@ -156,11 +155,8 @@ public abstract class CompressionMode {
     }
 
     @Override
-    public void compress(ByteBuffersDataInput buffersInput, DataOutput out) throws IOException {
-      final int len = (int) buffersInput.size();
-      byte[] bytes = new byte[len];
-      buffersInput.readBytes(bytes, 0, len);
-      LZ4.compress(bytes, 0, len, out, ht);
+    public void compress(byte[] bytes, int off, int len, DataOutput out) throws IOException {
+      LZ4.compress(bytes, off, len, out, ht);
     }
 
     @Override
@@ -178,11 +174,8 @@ public abstract class CompressionMode {
     }
 
     @Override
-    public void compress(ByteBuffersDataInput buffersInput, DataOutput out) throws IOException {
-      final int len = (int) buffersInput.size();
-      byte[] bytes = new byte[len];
-      buffersInput.readBytes(bytes, 0, len);
-      LZ4.compress(bytes, 0, len, out, ht);
+    public void compress(byte[] bytes, int off, int len, DataOutput out) throws IOException {
+      LZ4.compress(bytes, off, len, out, ht);
     }
 
     @Override
@@ -264,13 +257,9 @@ public abstract class CompressionMode {
     }
 
     @Override
-    public void compress(ByteBuffersDataInput buffersInput, DataOutput out) throws IOException {
-      final int len = (int) buffersInput.size();
-
-      byte[] bytes = new byte[len];
-      buffersInput.readBytes(bytes, 0, len);
+    public void compress(byte[] bytes, int off, int len, DataOutput out) throws IOException {
       compressor.reset();
-      compressor.setInput(bytes, 0, len);
+      compressor.setInput(bytes, off, len);
       compressor.finish();
 
       if (compressor.needsInput()) {