--- a/lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/compressing/Compressor.java
+++ b/lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/compressing/Compressor.java
@@ -14,11 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.codecs.compressing;
+package org.apache.lucene.backward_codecs.compressing;
 
 import java.io.Closeable;
 import java.io.IOException;
-import org.apache.lucene.store.ByteBuffersDataInput;
 import org.apache.lucene.store.DataOutput;
 
 /** A data compressor. */
@@ -32,6 +31,5 @@ public abstract class Compressor implements Closeable {
    * necessary information so that a {@link Decompressor} will know when to stop decompressing bytes
    * from the stream.
    */
-  public abstract void compress(ByteBuffersDataInput buffersInput, DataOutput out)
-      throws IOException;
+  public abstract void compress(byte[] bytes, int off, int len, DataOutput out) throws IOException;
 }