… lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene50/compressing/Lucene50CompressingStoredFieldsReader.java
… lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene50/compressing/Lucene50CompressingTermVectorsReader.java
… lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene87/DeflateWithPresetDictCompressionMode.java
… lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/compressing/AbstractTestCompressionMode.java
… lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/compressing/TestFastCompressionMode.java
… lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/compressing/TestFastDecompressionMode.java
… lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/compressing/TestHighCompressionMode.java
… lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene50/Lucene50RWStoredFieldsFormat.java
… lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene50/compressing/Lucene50CompressingStoredFieldsWriter.java
… lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene50/compressing/Lucene50CompressingTermVectorsWriter.java
… lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene50/compressing/Lucene50RWCompressingStoredFieldsFormat.java
… lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene50/compressing/Lucene50RWCompressingTermVectorsFormat.java
… lucene/core/src/java/org/apache/lucene/codecs/lucene90/compressing/Lucene90CompressingStoredFieldsWriter.java
… lucene/core/src/java/org/apache/lucene/codecs/lucene90/compressing/Lucene90CompressingTermVectorsWriter.java
… lucene/core/src/java/org/apache/lucene/index/SortingStoredFieldsConsumer.java
… lucene/core/src/java/org/apache/lucene/store/ByteArrayDataInput.java
… lucene/core/src/java/org/apache/lucene/store/ByteBuffersDataInput.java
… lucene/core/src/java/org/apache/lucene/store/ByteBuffersDataOutput.java
… lucene/core/src/test/org/apache/lucene/codecs/compressing/AbstractTestCompressionMode.java
… lucene/core/src/test/org/apache/lucene/codecs/compressing/TestDeflateWithPresetDictCompressionMode.java
… lucene/core/src/test/org/apache/lucene/codecs/compressing/TestLZ4WithPresetDictCompressionMode.java
… lucene/core/src/test/org/apache/lucene/codecs/lucene90/compressing/TestCompressingStoredFieldsFormat.java
… lucene/core/src/test/org/apache/lucene/store/TestByteBuffersDataInput.java
… lucene/test-framework/src/java/org/apache/lucene/tests/codecs/compressing/dummy/DummyCompressingCodec.java