@luyuncheng 2022-07-12
> Would it be possible to remove all `CompressionMode#compress` variants that take a `byte[]` now that you introduced a new method that takes a `ByteBuffersDataInput`?
> 
> Also maybe we should keep old codecs unmodified and only make this change to `Lucene90Codec` where it makes most sense?

Hi @jpountz Thanks for reviewing this code. 

I prefer keeping old codecs unmodified, because `CompressionMode#compress` is a public abstract method, if we change it with variants `ByteBuffersDataInput` we need to backport in many codecs, like [commits](https://github.com/apache/lucene/blob/382962f22df3ee3af3fb538b877c98d61a622ddb/lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene50/Lucene50StoredFieldsFormat.java).

And if we only using compress method with variants ByteBuffersDataInput in LUCENE90, we can not using abstract method `Compressor.compress`, when we want to use other compression mode.

Would it be possible to add a new method in Compressor, like following? it can keep the old codecs unmodified, and method with variants ByteBuffersDataInput only can be called in Lucene90Codec.

```
public abstract void compress(byte[] bytes, int off, int len, DataOutput out) throws IOException;

public void compress(CompositeByteBuf compositeByteBuf, DataOutput out) throws IOException;
```