@luyuncheng 2022-07-18
> Thanks for the iteration, the high-level idea looks good to me, I left some suggestions.

@jpountz Thanks for your nice suggestions. at latest commits [ebcbdd](https://github.com/apache/lucene/commit/457bab5abfbaa3aabebe99a97e645b4f623be15f)
1. change statement byteBuffers in Lucene90CompressingStoredFieldsWriter
2. move sliceOne logic into ByteBuffersDataInput level and it would determine whether use internal buffers or create a new buffer
3. delete DeflateWithPresetDictCompressionMode#doCompress method which param is byte[]
4. delete copyOneDoc-> readBytes which instance ByteArrayDataInput optimize. and i would create a new issue to reduce memory copy in copyOneDoc