@luyuncheng 2022-08-01
> Thanks for running these tests, let's remove `readNBytes` and switch to a shared byte[] that we call `DataInput#readBytes` on instead.

@jpountz Thanks a lot for you reviewing for the code. at https://github.com/apache/lucene/pull/987/commits/c0d31d3134653bf6009b798925fa2350e5f7ec9c i removed readNBytes and using shared byte[] buffer in `DeflateWithPresetDictCompressionMode`. 

I try to do more test about why `readNBytes` with a lot of memory copy, i think this is relevant to the:
1. `DeflateWithPresetCompressingCodec` chunk size: `1 << 18`
2. `LZ4WithPresetDictCompressionMode` NUM_SUB_BLOCKS: `10` 
3. `ByteBuffersDataOutput` BITS_PER_BLOCK

The `chunk size` is larger than `block size` which made `ByteBuffersDataInput` fragment always copy datas